fix: Add configmap generation to support omw model
As we are using oras to pull an image down from ecr at init time,
we cannot use the simple kustomize route of just setting image for the
oras versions. Added bespoke kustomize for seldondeployment env path
so we can pass the omw version as an env var.

Think it will be better to move to an external properties file, but
want to try with this initially to see if it works.

APP-2474